requirements.txt
spotipy>=2.25.0
mutagen>=1.47.0
rapidfuzz>=3.9.0
python-dotenv>=1.0.1
